check condition before rename file

+++ build.sh
@@ -1,3 +1,5 @@
 #!/usr/bin/bash
 go build .
-mv cobra-todo-cli todo-cli
+if [ -f "cobra-todo-cli" ]; then
+	mv cobra-todo-cli todo-cli
+fi
